+++ addOns/client/src/main/java/org/zaproxy/addon/client/ExtensionClientIntegration.java
@@ -45,11 +45,15 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebDriverException;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.control.Control.Mode;
 import org.parosproxy.paros.core.scanner.Plugin.AlertThreshold;
+import org.parosproxy.paros.db.Database;
+import org.parosproxy.paros.db.DatabaseException;
+import org.parosproxy.paros.db.DatabaseUnsupportedException;
 import org.parosproxy.paros.extension.Extension;
 import org.parosproxy.paros.extension.ExtensionAdaptor;
 import org.parosproxy.paros.extension.ExtensionHook;
@@ -69,6 +73,8 @@ import org.zaproxy.addon.client.internal.ClientSideDetails;
 import org.zaproxy.addon.client.internal.ReportedElement;
 import org.zaproxy.addon.client.internal.ReportedEvent;
 import org.zaproxy.addon.client.internal.ReportedObject;
+import org.zaproxy.addon.client.internal.db.ClientHistoryDao;
+import org.zaproxy.addon.client.internal.db.TableJdo;
 import org.zaproxy.addon.client.pscan.ClientPassiveScanController;
 import org.zaproxy.addon.client.pscan.ClientPassiveScanHelper;
 import org.zaproxy.addon.client.pscan.ClientPassiveScanRule;
@@ -153,6 +159,7 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
     private EventConsumer eventConsumer;
     private Event lastAjaxSpiderStartEvent;
     private static ImageIcon icon;
+    private TableJdo tableJdo;
 
     private ClientSpiderDialog spiderDialog;
     private ZapMenuItem menuItemCustomScan;
@@ -407,6 +414,18 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
                         eventConsumer, "org.zaproxy.zap.extension.spiderAjax.SpiderEventPublisher");
     }
 
+    @Override
+    public void postInstall() {
+        loadClientHistory();
+    }
+
+    private void loadClientHistory() {
+        if (clientHistoryTableModel != null) {
+            clientHistoryTableModel.clear();
+            clientHistoryTableModel.addReportedObjects(ClientHistoryDao.loadAll());
+        }
+    }
+
     public ClientOptions getClientParam() {
         if (clientParam == null) {
             clientParam = new ClientOptions();
@@ -494,6 +513,23 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
     @Override
     public void destroy() {
         this.spiderScanController.stopAllScans();
+        if (tableJdo != null) {
+            tableJdo.unload();
+        }
+    }
+
+    @Override
+    public boolean supportsDb(String type) {
+        return true;
+    }
+
+    @Override
+    public void databaseOpen(Database db) throws DatabaseException, DatabaseUnsupportedException {
+        try {
+            tableJdo = new TableJdo(db);
+        } catch (Exception e) {
+            LOGGER.warn(e.getMessage(), e);
+        }
     }
 
     public ClientNode getOrAddClientNode(String url, boolean visited, boolean storage) {
@@ -615,6 +651,7 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
             }
         }
         this.clientHistoryTableModel.addReportedObject(obj);
+        ClientHistoryDao.persist(obj);
         incPscanCount();
         this.passiveScanController
                 .getEnabledScanRules()
@@ -839,6 +876,10 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
             spiderScanController.stopAllScans();
             spiderScanController.reset();
 
+            if (api != null) {
+                api.clear();
+            }
+
             if (hasView()) {
                 getClientSpiderPanel().reset();
                 if (spiderDialog != null) {
@@ -855,9 +896,7 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
             if (clientDetailsPanel != null) {
                 clientDetailsPanel.clear();
             }
-            if (clientHistoryTableModel != null) {
-                clientHistoryTableModel.clear();
-            }
+            loadClientHistory();
             spiderScanController.reset();
 
             if (hasView()) {
@@ -927,6 +966,10 @@ public class ExtensionClientIntegration extends ExtensionAdaptor {
      */
     public void unregisterClientCallBack(ClientCallBackImplementor callback) {
         this.api.unregisterClientCallBack(callback);
+    }
+
+    public void browserClosing(WebDriver wd) {
+        this.api.browserClosing(wd);
     }
 
     private class ClientPassiveScanRuleProvider implements PassiveScanRuleProvider {