+++ addOns/webdrivers/webdrivermacos/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 
+## [191] - 2026-04-08
+### Changed
+- Update ChromeDriver to 147.0.7727.56.
+
 ## [190] - 2026-03-23
 ### Changed
 - Update ChromeDriver to 146.0.7680.165.
@@ -791,6 +795,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First release: Firefox v0.13.0 Chrome v2.27
 
+[191]: https://github.com/zaproxy/zap-extensions/releases/webdrivermacos-v191
 [190]: https://github.com/zaproxy/zap-extensions/releases/webdrivermacos-v190
 [189]: https://github.com/zaproxy/zap-extensions/releases/webdrivermacos-v189
 [188]: https://github.com/zaproxy/zap-extensions/releases/webdrivermacos-v188